+++ nl/lib/nl.rb
@@ -1,4 +1,5 @@
 require_relative 'nl/error'
+require_relative 'nl/bitfield32'
 require_relative 'nl/raw/wire'
 require_relative 'nl/structured_payload'
 require_relative 'nl/sub_message'